+++ controller/get.go
@@ -17,20 +17,32 @@ var timeoutRegexp = regexp.MustCompile(`(t\=\d+)\/?`)
 // <data block>
 // END
 func (self *Controller) Get(input []string) error {
+	var err error
 	cmd := parseGetCommand(input)
 
 	switch cmd.SubCommand {
 	case "", "open":
-		return self.get(cmd)
+		err = self.get(cmd)
 	case "close":
-		return self.getClose(cmd)
+		err = self.getClose(cmd)
+	case "close/open":
+		if err = self.getClose(cmd); err == nil {
+			err = self.get(cmd)
+		}
 	case "abort":
-		return self.getAbort(cmd)
+		err = self.getAbort(cmd)
 	case "peek":
-		return self.peek(cmd)
+		err = self.peek(cmd)
+	default:
+		err = errors.New("ERROR " + "Invalid command")
 	}
 
-	return errors.New("ERROR " + "Invalid command")
+	if err != nil {
+		return err
+	}
+	fmt.Fprint(self.rw.Writer, "END\r\n")
+	self.rw.Writer.Flush()
+	return nil
 }
 
 func (self *Controller) get(cmd *Command) error {
@@ -48,12 +60,10 @@ func (self *Controller) get(cmd *Command) error {
 		fmt.Fprintf(self.rw.Writer, "VALUE %s 0 %d\r\n", cmd.QueueName, len(item.Value))
 		fmt.Fprintf(self.rw.Writer, "%s\r\n", item.Value)
 	}
-	fmt.Fprint(self.rw.Writer, "END\r\n")
-	if cmd.SubCommand == "open" && len(item.Value) > 0 {
+	if strings.Contains(cmd.SubCommand, "open") && len(item.Value) > 0 {
 		self.setCurrentState(cmd, item)
 		q.AddOpenTransactions(1)
 	}
-	self.rw.Writer.Flush()
 	atomic.AddUint64(&self.repo.Stats.CmdGet, 1)
 	return nil
 }
@@ -69,15 +79,14 @@ func (self *Controller) getClose(cmd *Command) error {
 		self.setCurrentState(nil, nil)
 	}
 
-	fmt.Fprint(self.rw.Writer, "END\r\n")
-	self.rw.Writer.Flush()
 	return nil
 }
 
 func (self *Controller) getAbort(cmd *Command) error {
-	self.abort(cmd)
-	fmt.Fprint(self.rw.Writer, "END\r\n")
-	self.rw.Writer.Flush()
+	err := self.abort(cmd)
+	if err != nil {
+		return err
+	}
 	return nil
 }
 
@@ -111,8 +120,6 @@ func (self *Controller) peek(cmd *Command) error {
 		fmt.Fprintf(self.rw.Writer, "VALUE %s 0 %d\r\n", cmd.QueueName, len(item.Value))
 		fmt.Fprintf(self.rw.Writer, "%s\r\n", item.Value)
 	}
-	fmt.Fprint(self.rw.Writer, "END\r\n")
-	self.rw.Writer.Flush()
 	atomic.AddUint64(&self.repo.Stats.CmdGet, 1)
 	return nil
 }