+++ controller/get_test.go
@@ -7,6 +7,29 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func Test_parseGetCommand(t *testing.T) {
+	testCases := map[string]string{
+		"work":                         "",
+		"work/open":                    "open",
+		"work/close":                   "close",
+		"work/abort":                   "abort",
+		"work/peek":                    "peek",
+		"work/t=10":                    "",
+		"work/t=10/t=100/t=1234567890": "",
+		"work/t=10/open":               "open",
+		"work/open/t=10":               "open",
+		"work/close/open/t=10":         "close/open",
+		"work/close/t=10/open/abort":   "close/open/abort",
+	}
+
+	for input, subCommand := range testCases {
+		cmd := parseGetCommand([]string{"get", input})
+		assert.Equal(t, "get", cmd.Name, input)
+		assert.Equal(t, "work", cmd.QueueName, input)
+		assert.Equal(t, subCommand, cmd.SubCommand, input)
+	}
+}
+
 // Initialize queue 'test' with 1 item
 // get test = value
 // get test = empty
@@ -30,7 +53,7 @@ func Test_Get(t *testing.T) {
 	command := []string{"get", "test"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "VALUE test 0 10\r\n0123456789\r\nEND\r\n")
+	assert.Equal(t, "VALUE test 0 10\r\n0123456789\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -39,7 +62,7 @@ func Test_Get(t *testing.T) {
 	command = []string{"get", "test"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "END\r\n")
+	assert.Equal(t, "END\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -48,7 +71,7 @@ func Test_Get(t *testing.T) {
 	command = []string{"get", "test/close"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "END\r\n")
+	assert.Equal(t, "END\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -57,7 +80,7 @@ func Test_Get(t *testing.T) {
 	command = []string{"get", "test/close"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "END\r\n")
+	assert.Equal(t, "END\r\n", mockTCPConn.WriteBuffer.String())
 }
 
 // Initialize test queue with 4 items
@@ -90,14 +113,14 @@ func Test_GetOpen(t *testing.T) {
 	command := []string{"get", "test/open"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "VALUE test 0 1\r\n1\r\nEND\r\n")
+	assert.Equal(t, "VALUE test 0 1\r\n1\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
 	// get test = error
 	command = []string{"get", "test"}
 	err = controller.Get(command)
-	assert.Equal(t, err.Error(), "CLIENT_ERROR Close current item first")
+	assert.Equal(t, "CLIENT_ERROR Close current item first", err.Error())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -105,7 +128,7 @@ func Test_GetOpen(t *testing.T) {
 	command = []string{"get", "test/close"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "END\r\n")
+	assert.Equal(t, "END\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -113,7 +136,7 @@ func Test_GetOpen(t *testing.T) {
 	command = []string{"get", "test/open"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "VALUE test 0 1\r\n2\r\nEND\r\n")
+	assert.Equal(t, "VALUE test 0 1\r\n2\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 
 	// get test/open = error
 	command = []string{"get", "test/open"}
@@ -126,7 +149,7 @@ func Test_GetOpen(t *testing.T) {
 	command = []string{"get", "test/abort"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "END\r\n")
+	assert.Equal(t, "END\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -134,7 +157,7 @@ func Test_GetOpen(t *testing.T) {
 	command = []string{"get", "test/open"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "VALUE test 0 1\r\n2\r\nEND\r\n")
+	assert.Equal(t, "VALUE test 0 1\r\n2\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -142,7 +165,7 @@ func Test_GetOpen(t *testing.T) {
 	command = []string{"get", "test/peek"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "VALUE test 0 1\r\n3\r\nEND\r\n")
+	assert.Equal(t, "VALUE test 0 1\r\n3\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -150,7 +173,7 @@ func Test_GetOpen(t *testing.T) {
 	command = []string{"get", "test/close"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "END\r\n")
+	assert.Equal(t, "END\r\n", mockTCPConn.WriteBuffer.String())
 }
 
 // Initialize test queue with 2 items
@@ -177,7 +200,7 @@ func Test_GetOpen_Disconnect(t *testing.T) {
 	command := []string{"get", "test/open"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "VALUE test 0 1\r\n1\r\nEND\r\n")
+	assert.Equal(t, "VALUE test 0 1\r\n1\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 
 	mockTCPConn.WriteBuffer.Reset()
 
@@ -190,28 +213,80 @@ func Test_GetOpen_Disconnect(t *testing.T) {
 	command = []string{"get", "test"}
 	err = controller.Get(command)
 	assert.Nil(t, err)
-	assert.Equal(t, mockTCPConn.WriteBuffer.String(), "VALUE test 0 1\r\n1\r\nEND\r\n")
+	assert.Equal(t, "VALUE test 0 1\r\n1\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 }
 
-func Test_parseGetCommand(t *testing.T) {
-	testCases := map[string]string{
-		"work":                         "",
-		"work/open":                    "open",
-		"work/close":                   "close",
-		"work/abort":                   "abort",
-		"work/peek":                    "peek",
-		"work/t=10":                    "",
-		"work/t=10/t=100/t=1234567890": "",
-		"work/t=10/open":               "open",
-		"work/open/t=10":               "open",
-		"work/close/open/t=10":         "close/open",
-		"work/close/t=10/open/abort":   "close/open/abort",
-	}
+// Initialize test queue with 4 items
+// get test/close/open = value
+// get test = error
+// get test/t=10/close/open = value
+// get test/close/open/t=1000 = next value
+// FinishSession (disconnect)
+// get test/close/t=88/open = same value
+func Test_GetCloseOpen(t *testing.T) {
+	repo, err := repository.Initialize(dir)
+	defer repo.CloseAllQueues()
+	assert.Nil(t, err)
 
-	for input, subCommand := range testCases {
-		cmd := parseGetCommand([]string{"get", input})
-		assert.Equal(t, "get", cmd.Name, input)
-		assert.Equal(t, "work", cmd.QueueName, input)
-		assert.Equal(t, subCommand, cmd.SubCommand, input)
-	}
+	mockTCPConn := NewMockTCPConn()
+	controller := NewSession(mockTCPConn, repo)
+
+	repo.FlushQueue("test")
+	q, err := repo.GetQueue("test")
+	assert.Nil(t, err)
+
+	q.Enqueue([]byte("1"))
+	q.Enqueue([]byte("2"))
+	q.Enqueue([]byte("3"))
+	q.Enqueue([]byte("4"))
+
+	// get test/close/open = 1
+	command := []string{"get", "test/close/open"}
+	err = controller.Get(command)
+	assert.Nil(t, err)
+	assert.Equal(t, "VALUE test 0 1\r\n1\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
+
+	mockTCPConn.WriteBuffer.Reset()
+
+	// get test = error
+	command = []string{"get", "test"}
+	err = controller.Get(command)
+	assert.Equal(t, err.Error(), "CLIENT_ERROR Close current item first")
+
+	mockTCPConn.WriteBuffer.Reset()
+
+	// get test/abort = value
+	command = []string{"get", "test/abort"}
+	err = controller.Get(command)
+	assert.Nil(t, err)
+	assert.Equal(t, "END\r\n", mockTCPConn.WriteBuffer.String())
+
+	mockTCPConn.WriteBuffer.Reset()
+
+	// get test/t=10/close/open = value
+	command = []string{"get", "test/t=10/close/open"}
+	err = controller.Get(command)
+	assert.Nil(t, err)
+	assert.Equal(t, "VALUE test 0 1\r\n1\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
+
+	mockTCPConn.WriteBuffer.Reset()
+
+	// get test/close/open/t=1000 = next value
+	command = []string{"get", "test/close/open/t=1000"}
+	err = controller.Get(command)
+	assert.Nil(t, err)
+	assert.Equal(t, "VALUE test 0 1\r\n2\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
+
+	mockTCPConn.WriteBuffer.Reset()
+
+	controller.FinishSession()
+
+	mockTCPConn = NewMockTCPConn()
+	controller = NewSession(mockTCPConn, repo)
+
+	// get test = same value
+	command = []string{"get", "test/t=88/open"}
+	err = controller.Get(command)
+	assert.Nil(t, err)
+	assert.Equal(t, "VALUE test 0 1\r\n2\r\nEND\r\n", mockTCPConn.WriteBuffer.String())
 }